remove world_logger_test.go
